pyproject.toml
[project]
name = "mantra-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "psycopg>=3.2.10",
]
